requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.2.0
attrs==23.1.0
Automat==24.8.1
beautifulsoup4==4.12.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
cymem==2.0.8
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
filelock==3.15.4
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
gotrue==2.8.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.25.2
huggingface-hub==0.20.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
langcodes==3.4.0
language_data==1.2.0
lxml==4.9.3
marisa-trie==1.2.0
Markdown==3.5.2
MarkupSafe==2.1.4
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
networkx==3.3
nlp==0.4.0
nltk==3.9.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.12.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.0
parsel==1.8.1
pillow==10.2.0
playwright==1.46.0
postgrest==0.16.11
preshed==3.0.9
prompt-toolkit==3.0.40
Protego==0.3.1
protobuf==4.21.12
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
pyee==11.1.0
PyMySQL==1.1.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.2
queuelib==1.7.0
realtime==2.0.2
regex==2023.12.25
requests==2.25.1
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.4.0
scipy==1.12.0
Scrapy==2.11.0
scrapy-playwright==0.0.33
selenium==4.15.2
sentence-transformers==2.3.1
sentencepiece==0.1.99
service-identity==24.1.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.18
srsly==2.4.8
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
sympy==1.13.2
thinc==8.2.5
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.15.2
torch==2.4.1
tqdm==4.66.1
transformers==4.37.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
w3lib==2.1.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
websockets==12.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.9.10
zope.interface==7.0.3
